Refactor: mod and warn
